requirements.txt
Django==4.0.4
djangorestframework==3.13.1
django-cors-headers==3.11.0
mysqlclient==2.1.0
pymongo==4.1.1
dnspython==2.2.1
boto3==1.21.42